expose -> expose_service, ingress -> expose_ingress
