+++ deploy/operator/basic/role.yaml
@@ -64,12 +64,4 @@ rules:
   verbs:
     - get
     - list
-- apiGroups:
-    - networking.k8s.io
-    - extensions
-  resources:
-    - ingresses
-  verbs:
-    - get
-    - list
     - watch
