+++ pkg/deploy/spindeploy/expose_ingress/ingress_test.go
@@ -1,4 +1,4 @@
-package ingress
+package expose_ingress
 
 import (
 	"context"
