+++ pkg/deploy/spindeploy/expose_ingress/testdata/spinsvc_expose_ingress.yml
