+++ pkg/deploy/spindeploy/expose_service/testdata/input_service.yml
