+++ pkg/deploy/spindeploy/expose_service/testdata/output_service_lb.yml
