+++ pkg/deploy/spindeploy/expose_service/testdata/service.yml
