+++ pkg/deploy/spindeploy/expose_service/testdata/spinsvc_expose.yml
