+++ pkg/deploy/spindeploy/expose_service/testdata/spinsvc_minimal.yml
