+++ pkg/deploy/spindeploy/expose_service/transform_test.go
@@ -1,4 +1,4 @@
-package expose
+package expose_service
 
 import (
 	"context"
