+++ pkg/deploy/spindeploy/spindeploy.go
@@ -7,8 +7,8 @@ import (
 	"github.com/armory/spinnaker-operator/pkg/deploy"
 	"github.com/armory/spinnaker-operator/pkg/deploy/spindeploy/changedetector"
 	"github.com/armory/spinnaker-operator/pkg/deploy/spindeploy/config"
-	"github.com/armory/spinnaker-operator/pkg/deploy/spindeploy/expose"
-	"github.com/armory/spinnaker-operator/pkg/deploy/spindeploy/ingress"
+	"github.com/armory/spinnaker-operator/pkg/deploy/spindeploy/expose_ingress"
+	"github.com/armory/spinnaker-operator/pkg/deploy/spindeploy/expose_service"
 	"github.com/armory/spinnaker-operator/pkg/deploy/spindeploy/transformer"
 	"github.com/armory/spinnaker-operator/pkg/deploy/spindeploy/x509"
 	"github.com/go-logr/logr"
@@ -21,7 +21,7 @@ import (
 
 var DetectorGenerators = []changedetector.DetectorGenerator{
 	&config.ChangeDetectorGenerator{},
-	&expose.ChangeDetectorGenerator{},
+	&expose_service.ChangeDetectorGenerator{},
 	&x509.ChangeDetectorGenerator{},
 }
 
@@ -29,8 +29,8 @@ var TransformerGenerators = []transformer.Generator{
 	&transformer.OwnerTransformerGenerator{},
 	&transformer.NamedPortsTransformerGenerator{},
 	&transformer.TargetTransformerGenerator{},
-	&expose.TransformerGenerator{},
-	&ingress.TransformerGenerator{},
+	&expose_service.TransformerGenerator{},
+	&expose_ingress.TransformerGenerator{},
 	&transformer.ServerPortTransformerGenerator{},
 	&x509.X509TransformerGenerator{},
 	&transformer.AccountsTransformerGenerator{},
